Improve code documentation
